fixes, documentation, metadata and improves on CI/CD

+++ src/strings.rs
@@ -1,5 +1,5 @@
 pub fn count(text: String, substring: &str) -> usize {
-    text.matches(substring).count() as usize
+    text.matches(substring).count()
 }
